fix(docker): implement multi-stage build
Docker supports the concept of multi-stage builds. First you create
layers that contain all the development tools and build tools that are
required, then it takes the smallest possible image that can run the
application. That smallest possible image is what is pushed to the
registry.

Using this technique the image size becomes manageable, my results so
far:

- 1.2GB original biff Dockerfile
- 832MB removing the default-jre
- 600MB moving to Alpine as a base image
- 452MB using multi-stage buildes

Additional size reduction is possible through graalvm's native-image,
however I have not been able to build the biff starter with it.